Se añadio la condicion que no se puede usar la operacion % con 0, el TME no puede ser menor o igual a 0, y los numeros ingresados se pasaron a flotantes

+++ operacion.hpp
@@ -5,26 +5,26 @@
 #define OPERACION_HPP
 class Operacion{
     private:
-        int operando1;
-        int operando2;
+        float operando1;
+        float operando2;
         char operador;
-        int resultado;
+        float resultado;
     public:
         Operacion();
-        Operacion(int, int, char);
+        Operacion(float, float, char);
 
-        void setOperando1(int);
-        void setOperando2(int);
+        void setOperando1(float);
+        void setOperando2(float);
         void setOperador(char);
-        void setResultado(int);
+        void setResultado(float);
 
-        int getOperador1();
-        int getOperador2();
+        float getOperador1();
+        float getOperador2();
         char getOperador();
-        int getResultado();
+        float getResultado();
         std::string getOperacion();
 
-        int crearResultado(int, int, char);
+        float crearResultado(float, float, char);
         std::string toString();
 };
